changes to the example
